full board and clean up
