+++ constants/board-constants.ts
@@ -1,0 +1,14 @@
+import { Priority, Item, ItemStatus } from "../api/models";
+
+export const emptyItem: Item = {
+  id: "",
+  content: "",
+  title: "",
+  label: "",
+  user: "",
+  updateDateTime: "",
+  creationDateTime: "",
+  itemStatus: ItemStatus.BACKLOG,
+  priority: Priority.STANDARD,
+  position: 999,
+};
